chore: separate barrier fields
